Returns a consistent order for uncomparable literals
Before, if the two compared literals had typedValues
for which comparison was not supported by JavaScript,
the first argument was always considered as lower than the second.

Now, we compare first the literal typedValue types.
If they are equal, we proceed to the actual comparison.
Uf not, we compare the literals using their typedValue type name.
We do that instead of comparing directly the datatypes
to be sure we allow as many value comparisons as possible.